move pyg data class to its own module
